license_portal/requirements.txt
Django==3.2
psycopg2
djangorestframework==3.12.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
celery==5.2.1
redis==4.0.2
python-dotenv==0.19.2
python-dateutil==2.8.2
freezegun==1.2.2
django-cors-headers==3.13.0